backend/requirements.txt
aiohttp==3.9.3
anthropic==0.23.1
fastapi==0.110.1
httptools==0.6.1
ipykernel==6.29.3
pandas==2.2.1
pip-chill==1.0.3
suno-songs==0.4.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
